Many cleanup's. Adding in last error handling.
git-svn-id: http://cuda-waste.googlecode.com/svn/trunk@193 b05b5d53-d61a-d14b-5d44-457121efbe5e

Former-commit-id: fedfd1882866cbee864f9c5b376c09673c2fa873 [formerly 89b3fd895dd5387da1166fced71789933fe90fff] [formerly 02427661f6f0d903899588a9f3df393890780cdb [formerly 539f3c015e3ceeae124804a3d976683935931b0f]] [formerly bafbdb5e3c80c150069058e7d92c500b88298517 [formerly 133e63d973765897ec70787eb25fd71884e51436] [formerly 009ef7dcc6e79b7a69970c63a1b58e080ee388c3]] [formerly 5a549625a00638646c0db0332bd1fa3296a33c8c [formerly 3a4bd8d5c307c204909b2ba6fa52b6ddbf3dce08] [formerly 009ef7dcc6e79b7a69970c63a1b58e080ee388c3 [formerly 176cf141c45f87b74d6a9d573700984ab1e9c88e]]] [formerly 20ffde19518d7c4a86fc6fc7337591e5f3e8ab15 [formerly 71c9ad1af16f92502e654619befdf506f5f3282d] [formerly d3d3a1c93f29f609f9bfc9ef8d84b0e5deef544f [formerly 98dfb95a6da3fa6ae15bdca863773e3aa95833cf]] [formerly 5832279907b25d3b9c2bee3ff1ea4898e2270229 [formerly 2763de9e86984215bf73a89d007185ed33b8923d] [formerly f357ec6d35e9ced67e5baac5eb54029675aaa100]] [formerly 5a549625a00638646c0db0332bd1fa3296a33c8c [formerly 3a4bd8d5c307c204909b2ba6fa52b6ddbf3dce08] [formerly 009ef7dcc6e79b7a69970c63a1b58e080ee388c3 [formerly 176cf141c45f87b74d6a9d573700984ab1e9c88e]] [formerly 08eec6f991162d7873fdb2e898d62f6433a5ffd8 [formerly 182090a9bc611184053f494e1126c4a6f2366f32] [formerly f357ec6d35e9ced67e5baac5eb54029675aaa100 [formerly 450b7090b1670a209ba37b09e00818a99457a391]]]]]
Former-commit-id: 7b24bb51046315e536c80d69b4e1380a17780827 [formerly 22a7ab0ca413fb7756ce438e5c21a896e62afada] [formerly 7cf98a9343aef40e1e60f21f8059304366f52939 [formerly a482d4bc79d606cd1aff5dabbeab170f7b38bd68]] [formerly 64a0866c227c71c62ae53597a8e4194f61c067db [formerly 3ce44a3569050cb6195bba13bf51f36664497696] [formerly 08eec6f991162d7873fdb2e898d62f6433a5ffd8]] [formerly 32b0697c8cf63f75feaed600a8b174cb20e6c1b5 [formerly d478852e89072714bd508383776ed61fb5a22947] [formerly 80b4ef3f3d9e7397c271404243a99e499993bfe1 [formerly de5e28a6b4cd45177211286d6cdfd6dd2815e5a2]] [formerly 08eec6f991162d7873fdb2e898d62f6433a5ffd8 [formerly 182090a9bc611184053f494e1126c4a6f2366f32]]]
Former-commit-id: 35e62426bb0bbca58cb96e1d333c6fe346da08c1 [formerly f164a83793fb999debde5aa0f24c2e917059cdde] [formerly 3b3d961fb5822fe43200c6ad2cb74091a5f16a20 [formerly 9b57a4cb33f3583acf664133d29935e5051c0b71]] [formerly 32b0697c8cf63f75feaed600a8b174cb20e6c1b5 [formerly d478852e89072714bd508383776ed61fb5a22947] [formerly 80b4ef3f3d9e7397c271404243a99e499993bfe1 [formerly de5e28a6b4cd45177211286d6cdfd6dd2815e5a2]]]
Former-commit-id: 876dd222f4abb875266fd7a3958d833e54d1456c [formerly cfbf012f7f3628699497e7186cf4da94f6b63327] [formerly 80b4ef3f3d9e7397c271404243a99e499993bfe1]
Former-commit-id: 80dcdd25b3f323c6ace930fa4af88892250ba149 [formerly 9c13c23aa1c8febbd82360d47a372e6c70c9c2e5]
Former-commit-id: 83c4db179bdeb9ea7f44031c58be1967f5a0fd2f

+++ devices/tree.cpp
@@ -42,7 +42,7 @@ TREE * TREE::GetChild(int index)
         return this->children[index];
 }
 
-int TREE::GetChildCount()
+size_t TREE::GetChildCount()
 {
     return this->children.size();
 }
